PostTitleコンポーネントとPostImageコンポーネントを追加
