+++ app/my-jobs/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import SectionTitle from '../components/SectionTitle'
-import PostItems from '../components/PostItems'
+import PostItems from '../components/Post/PostItems'
 import { getListJob } from "../libs/microcms"
 
 //キャッシュを利用しない
